--- indexer/modules/hemera_ens/abi/0x3671ae578e63fdf66ad4f3e12cc0c0d71ac7510c.json
@@ -1,35 +1,0 @@
-{
-    "abi": [
-        {
-            "inputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "_ens",
-                    "type": "address"
-                }
-            ],
-            "stateMutability": "nonpayable",
-            "type": "constructor"
-        },
-        {
-            "inputs": [
-                {
-                    "internalType": "address[]",
-                    "name": "addresses",
-                    "type": "address[]"
-                }
-            ],
-            "name": "getNames",
-            "outputs": [
-                {
-                    "internalType": "string[]",
-                    "name": "r",
-                    "type": "string[]"
-                }
-            ],
-            "stateMutability": "view",
-            "type": "function"
-        }
-    ],
-    "address": "0x3671ae578e63fdf66ad4f3e12cc0c0d71ac7510c"
-}
