--- indexer/modules/hemera_ens/abi/0x4fe4e666be5752f1fdd210f4ab5de2cc26e3e0e8.json
@@ -1,143 +1,0 @@
-{
-    "abi": [
-        {
-            "inputs": [
-                {
-                    "internalType": "contract DNSSEC",
-                    "name": "_dnssec",
-                    "type": "address"
-                },
-                {
-                    "internalType": "contract ENS",
-                    "name": "_ens",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "constructor"
-        },
-        {
-            "anonymous": false,
-            "inputs": [
-                {
-                    "indexed": true,
-                    "internalType": "bytes32",
-                    "name": "node",
-                    "type": "bytes32"
-                },
-                {
-                    "indexed": true,
-                    "internalType": "address",
-                    "name": "owner",
-                    "type": "address"
-                },
-                {
-                    "indexed": false,
-                    "internalType": "bytes",
-                    "name": "dnsname",
-                    "type": "bytes"
-                }
-            ],
-            "name": "Claim",
-            "type": "event"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "internalType": "bytes",
-                    "name": "name",
-                    "type": "bytes"
-                },
-                {
-                    "internalType": "bytes",
-                    "name": "proof",
-                    "type": "bytes"
-                }
-            ],
-            "name": "claim",
-            "outputs": [],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "ens",
-            "outputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "oracle",
-            "outputs": [
-                {
-                    "internalType": "contract DNSSEC",
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "internalType": "bytes",
-                    "name": "name",
-                    "type": "bytes"
-                },
-                {
-                    "internalType": "bytes",
-                    "name": "input",
-                    "type": "bytes"
-                },
-                {
-                    "internalType": "bytes",
-                    "name": "proof",
-                    "type": "bytes"
-                }
-            ],
-            "name": "proveAndClaim",
-            "outputs": [],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [
-                {
-                    "internalType": "bytes4",
-                    "name": "interfaceID",
-                    "type": "bytes4"
-                }
-            ],
-            "name": "supportsInterface",
-            "outputs": [
-                {
-                    "internalType": "bool",
-                    "name": "",
-                    "type": "bool"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "pure",
-            "type": "function"
-        }
-    ],
-    "address": "0xa2f428617a523837d4adc81c67a296d42fd95e86"
-}
