--- indexer/modules/hemera_ens/abi/0x690f0581ececcf8389c223170778cd9d029606f2.json
@@ -1,22 +1,0 @@
-{
-    "abi": [
-        {
-            "inputs": [
-                {
-                    "internalType": "address",
-                    "name": "_masterCopy",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "constructor"
-        },
-        {
-            "payable": true,
-            "stateMutability": "payable",
-            "type": "fallback"
-        }
-    ],
-    "address": "0x690f0581ececcf8389c223170778cd9d029606f2"
-}
