--- indexer/modules/hemera_ens/abi/0x9062c0a6dbd6108336bcbe4593a3d1ce05512069.json
@@ -1,121 +1,0 @@
-{
-    "abi": [
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "name": "owner",
-                    "type": "address"
-                },
-                {
-                    "name": "resolver",
-                    "type": "address"
-                }
-            ],
-            "name": "claimWithResolver",
-            "outputs": [
-                {
-                    "name": "node",
-                    "type": "bytes32"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "name": "owner",
-                    "type": "address"
-                }
-            ],
-            "name": "claim",
-            "outputs": [
-                {
-                    "name": "node",
-                    "type": "bytes32"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "ens",
-            "outputs": [
-                {
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "defaultResolver",
-            "outputs": [
-                {
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [
-                {
-                    "name": "addr",
-                    "type": "address"
-                }
-            ],
-            "name": "node",
-            "outputs": [
-                {
-                    "name": "ret",
-                    "type": "bytes32"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "name": "name",
-                    "type": "string"
-                }
-            ],
-            "name": "setName",
-            "outputs": [
-                {
-                    "name": "node",
-                    "type": "bytes32"
-                }
-            ],
-            "payable": false,
-            "type": "function"
-        },
-        {
-            "inputs": [
-                {
-                    "name": "ensAddr",
-                    "type": "address"
-                },
-                {
-                    "name": "resolverAddr",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "type": "constructor"
-        }
-    ],
-    "address": "0x9062c0a6dbd6108336bcbe4593a3d1ce05512069"
-}
