--- indexer/modules/hemera_ens/abi/0xa2c122be93b0074270ebee7f6b7292c7deb45047.json
@@ -1,73 +1,0 @@
-{
-    "abi": [
-        {
-            "inputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "ensAddr",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "constructor"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "ens",
-            "outputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [
-                {
-                    "internalType": "bytes32",
-                    "name": "",
-                    "type": "bytes32"
-                }
-            ],
-            "name": "name",
-            "outputs": [
-                {
-                    "internalType": "string",
-                    "name": "",
-                    "type": "string"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "internalType": "bytes32",
-                    "name": "node",
-                    "type": "bytes32"
-                },
-                {
-                    "internalType": "string",
-                    "name": "_name",
-                    "type": "string"
-                }
-            ],
-            "name": "setName",
-            "outputs": [],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "function"
-        }
-    ],
-    "address": "0xa2c122be93b0074270ebee7f6b7292c7deb45047"
-}
