--- indexer/modules/hemera_ens/abi/0xff252725f6122a92551a5fa9a6b6bf10eb0be035.json
@@ -1,114 +1,0 @@
-{
-    "abi": [
-        {
-            "inputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "_ens",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "nonpayable",
-            "type": "constructor"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "BULK_RENEWAL_ID",
-            "outputs": [
-                {
-                    "internalType": "bytes4",
-                    "name": "",
-                    "type": "bytes4"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [],
-            "name": "ens",
-            "outputs": [
-                {
-                    "internalType": "contract ENS",
-                    "name": "",
-                    "type": "address"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": false,
-            "inputs": [
-                {
-                    "internalType": "string[]",
-                    "name": "names",
-                    "type": "string[]"
-                },
-                {
-                    "internalType": "uint256",
-                    "name": "duration",
-                    "type": "uint256"
-                }
-            ],
-            "name": "renewAll",
-            "outputs": [],
-            "payable": true,
-            "stateMutability": "payable",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [
-                {
-                    "internalType": "string[]",
-                    "name": "names",
-                    "type": "string[]"
-                },
-                {
-                    "internalType": "uint256",
-                    "name": "duration",
-                    "type": "uint256"
-                }
-            ],
-            "name": "rentPrice",
-            "outputs": [
-                {
-                    "internalType": "uint256",
-                    "name": "total",
-                    "type": "uint256"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "view",
-            "type": "function"
-        },
-        {
-            "constant": true,
-            "inputs": [
-                {
-                    "internalType": "bytes4",
-                    "name": "interfaceID",
-                    "type": "bytes4"
-                }
-            ],
-            "name": "supportsInterface",
-            "outputs": [
-                {
-                    "internalType": "bool",
-                    "name": "",
-                    "type": "bool"
-                }
-            ],
-            "payable": false,
-            "stateMutability": "pure",
-            "type": "function"
-        }
-    ],
-    "address": "0xff252725f6122a92551a5fa9a6b6bf10eb0be035"
-}
